This time I commited the right files. Also working on a prettier Marketplace interface.

+++ src/main/java/edu/gatech/gem5/game/ui/BuyBar.java
@@ -65,7 +65,7 @@ public class BuyBar extends HBox {
             @Override
             public void changed(ObservableValue<? extends Number> observable,
                     Number oldValue, Number newValue) {
-                lblQuantity.setText(newValue.toString());
+                lblQuantity.setText(((Integer) newValue.intValue()).toString());
             }
         });
     }
